@@ -14,18 +14,18 @@ description: A showcase sampling to list Movies from TMDB
 version: 1.1.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=3.2.0 <4.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  cupertino_icons: ^0.1.2
-  rxdart:
-  http:
-  firebase_core:
-  firebase_analytics:
-  firebase_crashlytics:
-  firebase_performance:
+  cupertino_icons: 1.0.8
+  rxdart: 0.28.0
+  http: 1.2.2
+  firebase_core: 3.4.1
+  firebase_analytics: 11.3.1
+  firebase_crashlytics: 4.1.1
+  firebase_performance: 0.10.0
 
 dev_dependencies:
   flutter_test:
@@ -53,22 +53,3 @@ flutter:
   # For details regarding adding assets from package dependencies, see
   # https://flutter.dev/assets-and-images/#from-packages
 
-  # To add custom fonts to your application, add a fonts section here,
-  # in this "flutter" section. Each entry in this list should have a
-  # "family" key with the font family name, and a "fonts" key with a
-  # list giving the asset and other descriptors for the font. For
-  # example:
-  # fonts:
-  #   - family: Schyler
-  #     fonts:
-  #       - asset: fonts/Schyler-Regular.ttf
-  #       - asset: fonts/Schyler-Italic.ttf
-  #         style: italic
-  #   - family: Trajan Pro
-  #     fonts:
-  #       - asset: fonts/TrajanPro.ttf
-  #       - asset: fonts/TrajanPro_Bold.ttf
-  #         weight: 700
-  #
-  # For details regarding fonts from package dependencies,
-  # see https://flutter.dev/custom-fonts/#from-packages
