@@ -1,42 +1,39 @@
 import 'package:MovieDirect/models/movie_item.dart';
 
 class MoviesList {
-  List<MovieItem> results;
-  int page;
-  int totalResults;
-  Dates dates;
-  int totalPages;
+   late List<MovieItem> results;
+  late int page;
+  late int totalResults;
+  late Dates dates;
+  late int totalPages;
 
-  MoviesList(
-      {this.results,
-      this.page,
-      this.totalResults,
-      this.dates,
-      this.totalPages});
+  MoviesList({
+    required this.results,
+    required this.page,
+    required this.totalResults,
+    required this.dates,
+    required this.totalPages,
+  });
 
   MoviesList.fromJson(Map<String, dynamic> json) {
+    results = <MovieItem>[]; // Initialize with an empty list
     if (json['results'] != null) {
-      results = new List<MovieItem>();
       json['results'].forEach((v) {
-        results.add(new MovieItem.fromJson(v));
+        results.add(MovieItem.fromJson(v));
       });
     }
-    page = json['page'];
-    totalResults = json['total_results'];
-    dates = json['dates'] != null ? new Dates.fromJson(json['dates']) : null;
-    totalPages = json['total_pages'];
+    page = json['page'] ?? 0; // Initialize with default value
+    totalResults = json['total_results'] ?? 0; // Initialize with default value
+    dates = (json['dates'] != null ? Dates.fromJson(json['dates']) : Dates(maximum: '', minimum: '')); // Initialize with default Dates
+    totalPages = json['total_pages'] ?? 0; // Initialize with default value
   }
 
   Map<String, dynamic> toJson() {
     final Map<String, dynamic> data = new Map<String, dynamic>();
-    if (this.results != null) {
-      data['results'] = this.results.map((v) => v.toJson()).toList();
-    }
+    data['results'] = this.results.map((v) => v.toJson()).toList();
     data['page'] = this.page;
     data['total_results'] = this.totalResults;
-    if (this.dates != null) {
-      data['dates'] = this.dates.toJson();
-    }
+    data['dates'] = this.dates.toJson();
     data['total_pages'] = this.totalPages;
     return data;
   }
@@ -46,12 +43,14 @@ class Dates {
   String maximum;
   String minimum;
 
-  Dates({this.maximum, this.minimum});
+  Dates({
+    required this.maximum,
+    required this.minimum,
+  });
 
-  Dates.fromJson(Map<String, dynamic> json) {
-    maximum = json['maximum'];
-    minimum = json['minimum'];
-  }
+  Dates.fromJson(Map<String, dynamic> json)
+      : maximum = json['maximum'] ?? '',
+        minimum = json['minimum'] ?? '';
 
   Map<String, dynamic> toJson() {
     final Map<String, dynamic> data = new Map<String, dynamic>();
